+++ project-extender/providers/base-provider/index.ts
@@ -1,0 +1,1 @@
+export * from "./BaseProvider"
