+++ project-extender/providers/jira-cloud-provider/JiraCloudProvider.ts
@@ -1,11 +1,15 @@
 /* eslint-disable max-classes-per-file */
 /* eslint-disable class-methods-use-this */
 import fetch from "cross-fetch"
-import { ProviderApi, ProviderCreator } from "./BaseProvider"
+import { ProviderApi, ProviderCreator } from "../base-provider"
 import { getAccessToken } from "./getAccessToken"
 
 class JiraCloudProvider implements ProviderApi {
-  async login(options: {
+  accessToken: string | undefined
+
+  async login({
+    oauthLoginOptions,
+  }: {
     oauthLoginOptions: {
       clientId: string
       clientSecret: string
@@ -13,20 +17,22 @@ class JiraCloudProvider implements ProviderApi {
       code: string
     }
   }) {
-    const accessToken = await getAccessToken(options.oauthLoginOptions)
-
-    await fetch("https://api.atlassian.com/oauth/token/accessible-resources", {
-      headers: {
-        Accept: "application/json",
-        Authorization: `Bearer ${accessToken}`,
-      },
-    }).then(async (response) => {
-      console.log(await response.json())
-    })
+    this.accessToken = await getAccessToken(oauthLoginOptions)
   }
 
   getProjects() {
     return [{ name: "name", key: "JIRAUSER10100" }]
+  }
+
+  async getResources() {
+    await fetch("https://api.atlassian.com/oauth/token/accessible-resources", {
+      headers: {
+        Accept: "application/json",
+        Authorization: `Bearer ${this.accessToken}`,
+      },
+    }).then(async (response) => {
+      console.log(await response.json())
+    })
   }
 }
