+++ project-extender/providers/jira-cloud-provider/index.ts
@@ -1,0 +1,1 @@
+export * from "./JiraCloudProvider"
