+++ project-extender/providers/jira-server-provider/index.ts
@@ -1,0 +1,1 @@
+export * from "./JiraServerProvider"
